Refactoring and other changes
