+++ samples/PleasantUI.Example/PleasantUIExampleApp.cs
@@ -1,6 +1,5 @@
 ﻿using Avalonia;
 using Avalonia.Controls;
-using Avalonia.Markup.Xaml;
 using PleasantUI.Core.Interfaces;
 using PleasantUI.Example.ViewModels;
