+++ src/PleasantUI/Controls/OptionsDisplayItem/OptionsDisplayItem.cs
@@ -1,6 +1,7 @@
 ﻿using System.Windows.Input;
 using Avalonia;
 using Avalonia.Controls;
+using Avalonia.Controls.Metadata;
 using Avalonia.Controls.Primitives;
 using Avalonia.Input;
 using Avalonia.Interactivity;
@@ -8,6 +9,7 @@ using Avalonia.Media;
 
 namespace PleasantUI.Controls;
 
+[TemplatePart("LayoutRoot", typeof(Border))]
 public class OptionsDisplayItem : TemplatedControl
 {
     private bool _isPressed;
