--- src/PleasantUI/Core/Exceptions/ApplicationNotInitializedException.cs
@@ -1,6 +1,0 @@
-﻿namespace PleasantUI.Core.Exceptions;
-
-/// <summary>
-/// Represents an exception that is thrown when the application is not initialized.
-/// </summary>
-public class ApplicationNotInitializedException(string? message) : Exception(message);
