+++ src/PleasantUI/Core/GenerationContexts/PleasantSettingsGenerationContext.cs
@@ -4,6 +4,4 @@ namespace PleasantUI.Core.GenerationContexts;
 
 [JsonSourceGenerationOptions(WriteIndented = true)]
 [JsonSerializable(typeof(PleasantSettings))]
-internal partial class PleasantSettingsGenerationContext : JsonSerializerContext
-{
-}
+internal partial class PleasantSettingsGenerationContext : JsonSerializerContext;
