--- src/PleasantUI/Styling/ControlThemes/OtherStyles/Transitions.axaml
@@ -1,24 +1,0 @@
-﻿<Styles xmlns="https://github.com/avaloniaui"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        x:ClassModifier="internal">
-    <Style Selector=":is(Button) > PathIcon,
-                     :is(Button) > StackPanel > PathIcon
-                     :is(Button) > DockPanel > PathIcon">
-        <Setter Property="Transitions">
-            <Transitions>
-                <TransformOperationsTransition Property="RenderTransform" Duration="0:0:.075" />
-            </Transitions>
-        </Setter>
-        <Setter Property="RenderTransform" Value="none" />
-    </Style>
-    <Style Selector=":is(Button):pointerover > PathIcon,
-                     :is(Button):pointerover > StackPanel > PathIcon,
-                     :is(Button):pointerover > DockPanel > PathIcon">
-        <Setter Property="RenderTransform" Value="scale(1.1)" />
-    </Style>
-    <Style Selector=":is(Button):pressed > PathIcon,
-                     :is(Button):pressed > StackPanel > PathIcon,
-                     :is(Button):pressed > DockPanel > PathIcon">
-        <Setter Property="RenderTransform" Value="scale(0.9)" />
-    </Style>
-</Styles>
